SQL и SQLite

+++ app/src/main/java/ru/netology/nmedia/dto/Post.kt
@@ -5,9 +5,9 @@ data class Post(
     val author: String,
     val content: String,
     val published: String,
-    val favorites: Int = 0,
-    var shares: Int = 0,
-    val views: Int = 0,
+    val favorites: Long = 0,
+    var shares: Long = 0,
+    val views: Long = 0,
     val favoritesByMe: Boolean,
     val sharesByMe: Boolean,
     val video: String?
